Sending VPN config to API
